should not exclude a group when has a known member

+++ src/main/java/br/com/unopay/api/uaa/service/GroupService.java
@@ -57,6 +57,9 @@ public class GroupService {
 
     public void delete(String id) {
         getById(id);
+        Page<UserDetail> members = findMembers(id, new UnovationPageRequest());
+        if(members.getContent() != null && !members.getContent().isEmpty())
+            throw UnovationExceptions.conflict().withErrors(GROUP_WITH_MEMBERS);
         repository.delete(id);
     }
